Update README.md
CURRENCY CONVERTER 
Welcome to the Currency Converter. This program enables a user to be able to basically convert the Kenyan Shilling(KES) to the provided currencies and the rates are as per the current world rates. There is a slot where one inputs the amount they want to convert from the Kenyan Shilling. The available currencies are then displayed and the user chooses from them. The conversion is then done as per the rates and the value is given to the user.
Slot 1. Enter Amount in KES: 40000
A display of the available currencies is given
Slot 2. Enter the target currency(the currency you want to convert to)USD
Amount in (target currency)USD: 263.504
Error Handling
If a character is put instead of a double or integer the system output will be "Wrong input. Try again" then the user will be able to try again.